Move Credential and CredentialRetriever to api package
