+++ jib-core/src/main/java/com/google/cloud/tools/jib/configuration/ImageConfiguration.java
@@ -16,8 +16,8 @@
 
 package com.google.cloud.tools.jib.configuration;
 
+import com.google.cloud.tools.jib.api.CredentialRetriever;
 import com.google.cloud.tools.jib.api.ImageReference;
-import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import java.util.List;
