+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/ContainerizerTest.java
@@ -17,7 +17,6 @@
 package com.google.cloud.tools.jib.api;
 
 import com.google.cloud.tools.jib.configuration.ImageConfiguration;
-import com.google.cloud.tools.jib.configuration.credentials.CredentialRetriever;
 import com.google.cloud.tools.jib.event.EventHandlers;
 import com.google.common.collect.ImmutableSet;
 import java.nio.file.Paths;
